requirements.txt
Flask==2.1.0
keras==2.8.0
nltk==3.7
numpy==1.21.0
tensorflow-cpu
gunicorn